requirements.txt
python-telegram-bot==13.7
google-generativeai==0.3.1
urllib3==1.26.15
